chore: move modal component
