+++ apps/web/src/components/Home/RecommendedProfiles.tsx
@@ -1,6 +1,5 @@
 import UserProfileShimmer from '@components/Shared/Shimmer/UserProfileShimmer';
 import UserProfile from '@components/Shared/UserProfile';
-import { Modal } from '@components/UI/Modal';
 import { DotsCircleHorizontalIcon, UsersIcon } from '@heroicons/react/outline';
 import { SparklesIcon } from '@heroicons/react/solid';
 import { Mixpanel } from '@lib/mixpanel';
@@ -10,7 +9,7 @@ import { useRecommendedProfilesQuery } from 'lens';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { FollowSource, MISCELLANEOUS } from 'src/tracking';
-import { Card, EmptyState, ErrorMessage } from 'ui';
+import { Card, EmptyState, ErrorMessage, Modal } from 'ui';
 
 import Suggested from './Suggested';
