+++ apps/web/src/components/Profile/NftGallery/Create.tsx
@@ -1,6 +1,5 @@
 import { useApolloClient } from '@apollo/client';
 import EmojiPicker from '@components/Shared/EmojiPicker';
-import { Modal } from '@components/UI/Modal';
 import { Spinner } from '@components/UI/Spinner';
 import { ChevronLeftIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
@@ -18,7 +17,7 @@ import React, { useState } from 'react';
 import { toast } from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { useNftGalleryStore } from 'src/store/nft-gallery';
-import { Button } from 'ui';
+import { Button, Modal } from 'ui';
 import trimify from 'utils/trimify';
 
 import Picker from './Picker';
