+++ apps/web/src/components/Profile/index.tsx
@@ -1,6 +1,5 @@
 import MetaTags from '@components/Common/MetaTags';
 import NftFeed from '@components/Nft/NftFeed';
-import { Modal } from '@components/UI/Modal';
 import { Mixpanel } from '@lib/mixpanel';
 import { APP_NAME, STATIC_IMAGES_URL } from 'data/constants';
 import { FeatureFlag } from 'data/feature-flags';
@@ -14,7 +13,7 @@ import Custom404 from 'src/pages/404';
 import Custom500 from 'src/pages/500';
 import { useAppStore } from 'src/store/app';
 import { PAGEVIEW } from 'src/tracking';
-import { GridItemEight, GridItemFour, GridLayout } from 'ui';
+import { GridItemEight, GridItemFour, GridLayout, Modal } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import isFeatureEnabled from 'utils/isFeatureEnabled';
