+++ apps/web/src/components/Publication/PublicationStats.tsx
@@ -1,13 +1,13 @@
 import Collectors from '@components/Shared/Modal/Collectors';
 import Likes from '@components/Shared/Modal/Likes';
 import Mirrors from '@components/Shared/Modal/Mirrors';
-import { Modal } from '@components/UI/Modal';
 import { CollectionIcon, HeartIcon, SwitchHorizontalIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
 import type { Publication } from 'lens';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { usePreferencesStore } from 'src/store/preferences';
+import { Modal } from 'ui';
 import nFormatter from 'utils/nFormatter';
 
 interface PublicationStatsProps {
