+++ apps/web/src/components/Shared/GlobalModals.tsx
@@ -1,10 +1,10 @@
 import Report from '@components/Shared/Modal/Report';
-import { Modal } from '@components/UI/Modal';
 import { ArrowCircleRightIcon, EmojiHappyIcon, ShieldCheckIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import type { FC } from 'react';
 import { useAuthStore } from 'src/store/auth';
 import { useGlobalModalStateStore } from 'src/store/modals';
+import { Modal } from 'ui';
 
 import Login from './Login';
 import Status from './Status';
