+++ apps/web/src/components/Shared/SuperFollow/index.tsx
@@ -1,4 +1,3 @@
-import { Modal } from '@components/UI/Modal';
 import { StarIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import { t } from '@lingui/macro';
@@ -9,7 +8,7 @@ import { useState } from 'react';
 import { useAppStore } from 'src/store/app';
 import { useAuthStore } from 'src/store/auth';
 import { PROFILE } from 'src/tracking';
-import { Button } from 'ui';
+import { Button, Modal } from 'ui';
 import formatHandle from 'utils/formatHandle';
 
 import Loader from '../Loader';
