+++ packages/ui/index.ts
@@ -9,3 +9,4 @@ export { GridItemEight, GridItemFour, GridLayout } from './GridLayout';
 export { default as HelpTooltip } from './HelpTooltip';
 export { Image } from './Image';
 export { Input } from './Input';
+export { Modal } from './Modal';
